create reverse_expand_weighted_test.go
